package name changed without using capital. when using databinding, capital package name can cause error.
