--- library/src/main/java/com/glide/slider/library/Transformers/DefaultTransformer.java
@@ -1,16 +1,0 @@
-package com.glide.slider.library.Transformers;
-
-import android.view.View;
-
-public class DefaultTransformer extends BaseTransformer {
-
-    @Override
-    protected void onTransform(View view, float position) {
-    }
-
-    @Override
-    public boolean isPagingEnabled() {
-        return true;
-    }
-
-}
