--- library/src/main/java/com/glide/slider/library/Transformers/StackTransformer.java
@@ -1,11 +1,0 @@
-package com.glide.slider.library.Transformers;
-
-import android.view.View;
-
-public class StackTransformer extends BaseTransformer {
-
-    @Override
-    protected void onTransform(View view, float position) {
-        view.setTranslationX(position < 0 ? 0f : -view.getWidth() * position);
-    }
-}
