--- library/src/main/java/com/glide/slider/library/Transformers/TabletTransformer.java
@@ -1,37 +1,0 @@
-package com.glide.slider.library.Transformers;
-
-import android.graphics.Camera;
-import android.graphics.Matrix;
-import android.view.View;
-
-public class TabletTransformer extends BaseTransformer {
-
-    private static final Matrix OFFSET_MATRIX = new Matrix();
-    private static final Camera OFFSET_CAMERA = new Camera();
-    private static final float[] OFFSET_TEMP_FLOAT = new float[2];
-
-    @Override
-    protected void onTransform(View view, float position) {
-        final float rotation = (position < 0 ? 30f : -30f) * Math.abs(position);
-
-        view.setTranslationX(getOffsetXForRotation(rotation, view.getWidth(), view.getHeight()));
-        view.setPivotX(view.getWidth() * 0.5f);
-        view.setPivotY(0);
-        view.setRotationY(rotation);
-    }
-
-    private static float getOffsetXForRotation(float degrees, int width, int height) {
-        OFFSET_MATRIX.reset();
-        OFFSET_CAMERA.save();
-        OFFSET_CAMERA.rotateY(Math.abs(degrees));
-        OFFSET_CAMERA.getMatrix(OFFSET_MATRIX);
-        OFFSET_CAMERA.restore();
-
-        OFFSET_MATRIX.preTranslate(-width * 0.5f, -height * 0.5f);
-        OFFSET_MATRIX.postTranslate(width * 0.5f, height * 0.5f);
-        OFFSET_TEMP_FLOAT[0] = width;
-        OFFSET_TEMP_FLOAT[1] = height;
-        OFFSET_MATRIX.mapPoints(OFFSET_TEMP_FLOAT);
-        return (width - OFFSET_TEMP_FLOAT[0]) * (degrees > 0.0f ? 1.0f : -1.0f);
-    }
-}
