+++ library/src/main/java/com/glide/slider/library/sliderTypes/TextSliderView.java
@@ -1,0 +1,28 @@
+package com.glide.slider.library.sliderTypes;
+
+import android.content.Context;
+import android.support.v7.widget.AppCompatImageView;
+import android.support.v7.widget.AppCompatTextView;
+import android.view.LayoutInflater;
+import android.view.View;
+
+import com.glide.slider.library.R;
+
+/**
+ * This is a slider with a description TextView.
+ */
+public class TextSliderView extends BaseSliderView {
+    public TextSliderView(Context context) {
+        super(context);
+    }
+
+    @Override
+    public View getView() {
+        View v = LayoutInflater.from(getContext()).inflate(R.layout.render_type_text, null);
+        AppCompatImageView target = v.findViewById(R.id.glide_slider_image);
+        AppCompatTextView description = v.findViewById(R.id.glide_slider_description);
+        description.setText(getDescription());
+        bindEventAndShow(v, target);
+        return v;
+    }
+}
