+++ library/src/main/java/com/glide/slider/library/transformers/AccordionTransformer.java
@@ -1,0 +1,16 @@
+package com.glide.slider.library.transformers;
+
+/**
+ * Created by daimajia on 14-5-29.
+ */
+
+import android.view.View;
+
+public class AccordionTransformer extends BaseTransformer {
+
+    @Override
+    protected void onTransform(View view, float position) {
+        view.setPivotX(position < 0 ? 0 : view.getWidth());
+        view.setScaleX(position < 0 ? 1f + position : 1f - position);
+    }
+}
