+++ library/src/main/java/com/glide/slider/library/transformers/BackgroundToForegroundTransformer.java
@@ -1,0 +1,23 @@
+package com.glide.slider.library.transformers;
+
+import android.view.View;
+
+public class BackgroundToForegroundTransformer extends BaseTransformer {
+
+    @Override
+    protected void onTransform(View view, float position) {
+        final float height = view.getHeight();
+        final float width = view.getWidth();
+        final float scale = min(position < 0 ? 1f : Math.abs(1f - position), 0.5f);
+
+        view.setScaleX(scale);
+        view.setScaleY(scale);
+        view.setPivotX(width * 0.5f);
+        view.setPivotY(height * 0.5f);
+        view.setTranslationX(position < 0 ? width * position : -width * position * 0.25f);
+    }
+
+    private static final float min(float val, float min) {
+        return val < min ? min : val;
+    }
+}
