+++ library/src/main/java/com/glide/slider/library/transformers/DefaultTransformer.java
@@ -1,0 +1,16 @@
+package com.glide.slider.library.transformers;
+
+import android.view.View;
+
+public class DefaultTransformer extends BaseTransformer {
+
+    @Override
+    protected void onTransform(View view, float position) {
+    }
+
+    @Override
+    public boolean isPagingEnabled() {
+        return true;
+    }
+
+}
